requirements.txt
playwright>=1.40.0
python-telegram-bot>=21.0
apscheduler>=3.10.0
sqlcipher3>=0.5.0
pyyaml>=6.0
